Added hierarchy to Editor

+++ src/main/java/nl/framegengine/engine/panels/InspectorPanel.java
@@ -5,7 +5,7 @@ import nl.framegengine.engine.EditorPanel;
 
 public class InspectorPanel extends EditorPanel {
 
-    public InspectorPanel(float posX, float posY, float sizeX, float sizeY) {
+    public InspectorPanel(int posX, int posY, int sizeX, int sizeY) {
         super(posX, posY, sizeX, sizeY);
     }
